+++ NEWS.md
@@ -1,3 +1,12 @@
+# eatATA 0.9.1
+
+* `GLPK` and `lpSolve` solvers are now supported
+* `useSolver()` function implemented, which calls solver APIs directly
+* `inspectSolution()` and `appendSolution()` functions, which use `useSovler()` output
+* `prepareConstraints()` and `processGurobiOutput()` retired
+* `calculateIIF()` calculates item information function
+* `calculateExpectedRT()` calculates expected response times
+
 # eatATA 0.7.0
 
 * Initial CRAN release.